Move mixins to separate folders
