+++ src/clauses/mixins/clauses/WithReturn.ts
@@ -17,9 +17,9 @@
  * limitations under the License.
  */
 
-import type { ProjectionColumn } from "../sub-clauses/Projection";
-import { Return } from "../Return";
-import { ClauseMixin } from "./ClauseMixin";
+import type { ProjectionColumn } from "../../sub-clauses/Projection";
+import { Return } from "../../Return";
+import { ClauseMixin } from "../ClauseMixin";
 
 export abstract class WithReturn extends ClauseMixin {
     protected returnStatement: Return | undefined;
